delineate DocumentPreview from DocumentEditor
