+++ src/components/Firestore/DocumentPreview/index.tsx
@@ -1,0 +1,77 @@
+/**
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React, { useState } from 'react';
+import { firestore } from 'firebase';
+import { List, ListItem } from '@rmwc/list';
+import { Button } from '@rmwc/button';
+import { useDocumentData } from 'react-firebase-hooks/firestore';
+
+import { updateField } from './api';
+import FieldPreview from './FieldPreview';
+import InlineEditor from './InlineEditor';
+import { FirestoreMap } from '../models';
+import { DocumentProvider } from './store';
+import { isMap } from '../utils';
+
+export interface Props {
+  reference: firestore.DocumentReference;
+}
+
+const DocumentPreview: React.FC<Props> = ({ reference }) => {
+  const [data, loading, error] = useDocumentData(reference);
+  const [isAddingField, setIsAddingField] = useState(false);
+
+  if (loading) return <div>Loading document...</div>;
+  if (error) return <div>Error retrieving document</div>;
+
+  return (
+    <>
+      <DocumentProvider value={data as FirestoreMap}>
+        <List dense className="List-Actions">
+          {/* Actions */}
+          <ListItem
+            tag={props => (
+              <Button dense label="Add field" icon="add" {...props} />
+            )}
+            onClick={() => setIsAddingField(true)}
+          ></ListItem>
+
+          {isAddingField && (
+            <InlineEditor
+              value={{ '': '' }}
+              path={[]}
+              onCancel={() => {
+                setIsAddingField(false);
+              }}
+              onSave={(key, value) => {
+                updateField(reference, data as FirestoreMap, [key], value[key]);
+                setIsAddingField(false);
+              }}
+            />
+          )}
+
+          {isMap(data) &&
+            Object.keys(data).map(name => (
+              <FieldPreview key={name} path={[name]} documentRef={reference} />
+            ))}
+        </List>
+      </DocumentProvider>
+    </>
+  );
+};
+
+export default DocumentPreview;
